refactor: move custom textformfield to own widget and rename
