+++ lib/screens/settings_screen/sections/pomodoro_settings_section.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:orodomop/providers/settings_provider.dart';
+import 'package:orodomop/screens/settings_screen/widgets/settings_duration_text_form.dart';
 import 'package:orodomop/screens/settings_screen/widgets/settings_item.dart';
 import 'package:orodomop/screens/settings_screen/widgets/settings_section_title.dart';
 import 'package:provider/provider.dart';
@@ -22,7 +23,7 @@ class PomodoroSettingsSection extends StatelessWidget {
             SettingsSectionTitle(title: "Pomodoro"),
             SettingsItem(
               text: "Focus duration (m)",
-              widget: PomodoroDurationTextForm(
+              widget: SettingsDurationTextForm(
                 seconds: model.focusDuration,
                 onSubmit: ({required seconds}) {
                   model.focusDuration = seconds;
@@ -32,7 +33,7 @@ class PomodoroSettingsSection extends StatelessWidget {
 
             SettingsItem(
               text: "Break duration (m)",
-              widget: PomodoroDurationTextForm(
+              widget: SettingsDurationTextForm(
                 seconds: model.breakDuration,
                 onSubmit: ({required seconds}) {
                   model.breakDuration = seconds;
@@ -41,68 +42,6 @@ class PomodoroSettingsSection extends StatelessWidget {
             ),
           ],
         );
-      },
-    );
-  }
-}
-
-class PomodoroDurationTextForm extends StatefulWidget {
-  const PomodoroDurationTextForm({
-    super.key,
-    required this.seconds,
-    required this.onSubmit,
-  });
-
-  final int seconds;
-  final Function({required int seconds}) onSubmit;
-
-  @override
-  State<PomodoroDurationTextForm> createState() =>
-      _PomodoroDurationTextFormState();
-}
-
-class _PomodoroDurationTextFormState extends State<PomodoroDurationTextForm> {
-  late TextEditingController _controller;
-  late FocusNode _focusNode;
-  late String originalValue;
-
-  @override
-  void initState() {
-    super.initState();
-    originalValue = (widget.seconds ~/ 60).toString();
-    _controller = TextEditingController(text: originalValue);
-    _focusNode = FocusNode();
-
-    _focusNode.addListener(() {
-      if (!_focusNode.hasFocus) {
-        _controller.text = originalValue;
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      controller: _controller,
-      focusNode: _focusNode,
-      keyboardType: TextInputType.number,
-      onFieldSubmitted: (value) {
-        int? minutes = int.tryParse(value);
-
-        if (minutes != null && minutes > 0) {
-          widget.onSubmit(seconds: minutes * 60);
-          originalValue = minutes.toString();
-        } else {
-          _controller.text = originalValue;
-        }
-
-        _focusNode.unfocus();
       },
     );
   }
